Implement an extensible formatter system.
Can register formatters and then call the right one depending on which
citation format you want to output. The eventual plan for this is to
enable add-on formatters that aren't included in core Ruby CFF, or
formatters that are only enabled if certain other dependencies are
available.
