+++ lib/cff/formatters/formatter.rb
@@ -18,74 +18,80 @@ require 'date'
 
 ##
 module CFF
-  # Formatter base class
-  class Formatter # :nodoc:
-    STATUS_TEXT_MAP = {
-      'advance-online' => 'Advance online publication',
-      'in-preparation' => 'Manuscript in preparation.',
-      'submitted' => 'Manuscript submitted for publication.'
-    }.freeze
+  module Formatters # :nodoc:
+    # Formatter base class
+    class Formatter # :nodoc:
+      STATUS_TEXT_MAP = {
+        'advance-online' => 'Advance online publication',
+        'in-preparation' => 'Manuscript in preparation.',
+        'submitted' => 'Manuscript submitted for publication.'
+      }.freeze
 
-    def self.select_and_check_model(model, preferred_citation)
-      if preferred_citation && model.preferred_citation.is_a?(Reference)
-        model = model.preferred_citation
+      def self.label
+        @label ||= name.split('::')[-1]
       end
 
-      # Safe to assume valid `Index`s and `Reference`s will have these fields.
-      model.authors.empty? || model.title.empty? ? nil : model
-    end
+      def self.select_and_check_model(model, preferred_citation)
+        if preferred_citation && model.preferred_citation.is_a?(Reference)
+          model = model.preferred_citation
+        end
 
-    def self.initials(name)
-      name.split.map { |part| part[0].capitalize }.join('. ')
-    end
-
-    def self.note_from_model(model)
-      STATUS_TEXT_MAP[model.status]
-    end
-
-    # Prefer `repository_code` over `url`
-    def self.url(model)
-      model.repository_code.empty? ? model.url : model.repository_code
-    end
-
-    def self.month_and_year_from_model(model) # rubocop:disable Metrics
-      return ['', 'in press'] if model.respond_to?(:status) && model.status == 'in-press'
-      if model.respond_to?(:year) && !model.year.to_s.empty?
-        return [model.month, model.year].map(&:to_s)
+        # Safe to assume valid `Index`s and `Reference`s will have these fields.
+        model.authors.empty? || model.title.empty? ? nil : model
       end
 
-      date = month_and_year_from_date(model.date_released)
-      if date == ['', ''] && model.respond_to?(:date_published)
-        date = month_and_year_from_date(model.date_published)
+      def self.initials(name)
+        name.split.map { |part| part[0].capitalize }.join('. ')
       end
-      date
-    end
 
-    def self.month_and_year_from_date(value)
-      if value.is_a?(Date)
-        [value.month, value.year].map(&:to_s)
-      else
-        begin
-          date = Date.parse(value.to_s)
-          [date.month, date.year].map(&:to_s)
-        rescue ArgumentError
-          ['', '']
+      def self.note_from_model(model)
+        STATUS_TEXT_MAP[model.status]
+      end
+
+      # Prefer `repository_code` over `url`
+      def self.url(model)
+        model.repository_code.empty? ? model.url : model.repository_code
+      end
+
+      def self.month_and_year_from_model(model) # rubocop:disable Metrics
+        return ['', 'in press'] if model.respond_to?(:status) && model.status == 'in-press'
+        if model.respond_to?(:year) && !model.year.to_s.empty?
+          return [model.month, model.year].map(&:to_s)
+        end
+
+        date = month_and_year_from_date(model.date_released)
+        if date == ['', ''] && model.respond_to?(:date_published)
+          date = month_and_year_from_date(model.date_published)
+        end
+        date
+      end
+
+      def self.month_and_year_from_date(value)
+        if value.is_a?(Date)
+          [value.month, value.year].map(&:to_s)
+        else
+          begin
+            date = Date.parse(value.to_s)
+            [date.month, date.year].map(&:to_s)
+          rescue ArgumentError
+            ['', '']
+          end
         end
       end
-    end
 
-    # CFF 'pages' is the number of pages, which has no equivalent in BibTeX
-    # or APA. References: https://www.bibtex.com/f/pages-field/,
-    # https://apastyle.apa.org/style-grammar-guidelines/references/examples
-    def self.pages_from_model(model, dash: '--')
-      return '' if !model.respond_to?(:start) || model.start.to_s.empty?
+      # CFF 'pages' is the number of pages, which has no equivalent in BibTeX
+      # or APA. References: https://www.bibtex.com/f/pages-field/,
+      # https://apastyle.apa.org/style-grammar-guidelines/references/examples
+      def self.pages_from_model(model, dash: '--')
+        return '' if !model.respond_to?(:start) || model.start.to_s.empty?
 
-      start = model.start.to_s
-      finish = model.end.to_s
-      if finish.empty?
-        start
-      else
-        start == finish ? start : "#{start}#{dash}#{finish}"
+        start = model.start.to_s
+        finish = model.end.to_s
+        if finish.empty?
+          start
+        else
+          start == finish ? start : "#{start}#{dash}#{finish}"
+        end
       end
     end
   end
