+++ lib/cff/index.rb
@@ -23,6 +23,7 @@ require_relative 'person'
 require_relative 'reference'
 require_relative 'schema'
 require_relative 'validatable'
+require_relative 'formatters'
 
 require 'yaml'
 
@@ -117,6 +118,28 @@ module CFF
     end
 
     # :call-seq:
+    #   citation(format, preferred_citation: true) -> String
+    #
+    # Output this Index in the specified format. Setting
+    # `preferred_citation: true` will honour the `preferred_citation` field in
+    # the index if one is present (default).
+    #
+    # `format` can be supplied as a String or a Symbol.
+    #
+    # Formats that are built-in to Ruby CFF are:
+    #
+    # * APAlike (e.g. `:apalike`, `'apalike'` or `'APAlike'`)
+    # * BibTeX (e.g. `:bibtex`, `'bibtex'` or `'BibTeX'`)
+    #
+    # *Note:* This method assumes that this Index is valid when called.
+    def citation(format, preferred_citation: true)
+      formatter = Formatters.formatter_for(format)
+      return '' if formatter.nil?
+
+      formatter.format(model: self, preferred_citation: preferred_citation)
+    end
+
+    # :call-seq:
     #   type = type
     #
     # Sets the type of this CFF Index. The type is currently restricted to one
@@ -140,9 +163,7 @@ module CFF
     #
     # *Note:* This method assumes that this Index is valid when called.
     def to_apalike(preferred_citation: true)
-      CFF::ApaFormatter.format(
-        model: self, preferred_citation: preferred_citation
-      )
+      citation(:apalike, preferred_citation: preferred_citation)
     end
 
     # :call-seq:
@@ -154,9 +175,7 @@ module CFF
     #
     # *Note:* This method assumes that this Index is valid when called.
     def to_bibtex(preferred_citation: true)
-      CFF::BibtexFormatter.format(
-        model: self, preferred_citation: preferred_citation
-      )
+      citation(:bibtex, preferred_citation: preferred_citation)
     end
 
     private
