+++ test/formatters/cff_bibtex_test.rb
@@ -2,7 +2,7 @@
 
 require_relative '../test_helper'
 
-require 'cff/formatters/bibtex_formatter'
+require 'cff/formatters/bibtex'
 require 'cff/file'
 
 class CFFBibtexFormatterTest < Minitest::Test
@@ -21,6 +21,10 @@ class CFFBibtexFormatterTest < Minitest::Test
     end
   end
 
+  def test_formatter_label
+    assert_equal('BibTeX', CFF::Formatters::BibTeX.label)
+  end
+
   def test_can_tolerate_invalid_file
     cff = CFF::Index.new(nil)
     assert_nil cff.to_bibtex
@@ -28,35 +32,35 @@ class CFFBibtexFormatterTest < Minitest::Test
 
   def test_bibtex_type
     index = ::CFF::Index.new('Title')
-    assert_equal('software', ::CFF::BibtexFormatter.bibtex_type(index))
+    assert_equal('software', CFF::Formatters::BibTeX.bibtex_type(index))
 
     index.type = 'software'
-    assert_equal('software', ::CFF::BibtexFormatter.bibtex_type(index))
+    assert_equal('software', CFF::Formatters::BibTeX.bibtex_type(index))
 
     index.type = 'dataset'
-    assert_equal('misc', ::CFF::BibtexFormatter.bibtex_type(index))
+    assert_equal('misc', CFF::Formatters::BibTeX.bibtex_type(index))
 
     ref = ::CFF::Reference.new('Title')
-    assert_equal('misc', ::CFF::BibtexFormatter.bibtex_type(ref))
+    assert_equal('misc', CFF::Formatters::BibTeX.bibtex_type(ref))
 
     ref.type = 'newspaper-article'
-    assert_equal('article', ::CFF::BibtexFormatter.bibtex_type(ref))
+    assert_equal('article', CFF::Formatters::BibTeX.bibtex_type(ref))
 
     ref.type = 'conference'
-    assert_equal('proceedings', ::CFF::BibtexFormatter.bibtex_type(ref))
+    assert_equal('proceedings', CFF::Formatters::BibTeX.bibtex_type(ref))
 
     ref.type = 'conference-paper'
-    assert_equal('inproceedings', ::CFF::BibtexFormatter.bibtex_type(ref))
+    assert_equal('inproceedings', CFF::Formatters::BibTeX.bibtex_type(ref))
 
     ref.type = 'proceedings'
-    assert_equal('proceedings', ::CFF::BibtexFormatter.bibtex_type(ref))
+    assert_equal('proceedings', CFF::Formatters::BibTeX.bibtex_type(ref))
 
     ref.type = 'pamphlet'
-    assert_equal('booklet', ::CFF::BibtexFormatter.bibtex_type(ref))
+    assert_equal('booklet', CFF::Formatters::BibTeX.bibtex_type(ref))
 
     ['article', 'book', 'manual', 'unpublished'].each do |type|
       ref.type = type
-      assert_equal(type, ::CFF::BibtexFormatter.bibtex_type(ref))
+      assert_equal(type, CFF::Formatters::BibTeX.bibtex_type(ref))
     end
   end
 
@@ -87,7 +91,7 @@ class CFFBibtexFormatterTest < Minitest::Test
         'Solskjaer_My_Strasse_2021'
       ]
     ].each do |fields, reference|
-      assert_equal(reference, ::CFF::BibtexFormatter.generate_citekey(fields))
+      assert_equal(reference, CFF::Formatters::BibTeX.generate_citekey(fields))
     end
   end
 end
